Endpoint /videos using media-api

+++ src/resources/comments.py
@@ -14,12 +14,16 @@ class Comments(Resource):
     @check_token
     def get(self, comment_id):
         schema = CommentSchema()
-        comment = Comment.objects(id=comment_id).first_or_404()
+        comment = Comment.objects(id=comment_id).first()
+        if comment is None:
+            return response_error(HTTPStatus.NOT_FOUND, "Comment not found")
         return make_response(schema.dump(comment), HTTPStatus.OK)
 
     @check_token
     def delete(self, comment_id):
-        comment = Comment.objects(id=comment_id).first_or_404()
+        comment = Comment.objects(id=comment_id).first()
+        if comment is None:
+            return response_error(HTTPStatus.NOT_FOUND, "Comment not found")
         if comment.user != g.session_username:
             return response_error(HTTPStatus.FORBIDDEN, str("Forbidden"))
         comment.delete()
@@ -42,9 +46,10 @@ class CommentsList(Resource):
             comment = schema.load(request.get_json(force=True))
         except ValidationError as e:
             return response_error(HTTPStatus.BAD_REQUEST, str(e.normalized_messages()))
-        Video.objects(id=comment.video.id).first_or_404()
-        if comment.parent:
-            Comment.objects(id=comment.parent.id, video=comment.video.id).first_or_404()
+        if Video.objects(id=comment.video.id).first() is None:
+            return response_error(HTTPStatus.NOT_FOUND, "Video not found")
+        if comment.parent and Comment.objects(id=comment.parent.id, video=comment.video.id).first() is None:
+            return response_error(HTTPStatus.NOT_FOUND, "Parent comment not found")
         now = datetime.datetime.utcnow()
         comment.user = g.session_username
         comment.date_created = now
